feat: extract func from filteralpha to loansearch
